Split source into three files
