+++ constants.go
@@ -1,0 +1,8 @@
+package main
+
+const VirusMarkOffset int = 3
+const VirusMarkEndOffset int = 5
+const VirusGenerationOffset int = 5
+const OriginalCodeOffset int = 0x34f
+
+const VirusNotFound int = -1
